de-indent correctly after typing 'fi'
Regression from ef2e050 "Better indenting for nested if sections" which
changed to using searchpair() instead of search().

When typing "fi", the cursor is after the start of the \<fi\> match, so
vim un-indents back to the next outer if/fi block rather than the
expected block.

The help for searchpair() suggests "When searching backwards and {end} is
more than one character, it may be useful to put "\zs" at the end of the
pattern, so that when the cursor is inside a match with the end it finds
the matching start." Follow this suggestion to restore the expected
behavior.

Fixes: https://github.com/chrisbra/vim-sh-indent/issues/15