Check for if keywords for zsh should not apply to comments
Currently, the check for zsh special constructs did not check that the
line does not begin with a '#' so commented lines would be indented
additionally if they e.g. contained an `if`.

This is certainly wrong and inconsistent with how the other part of the
shell indent script is working. So add the `^\s*` to the front of the
statements.

Suggested by Izaak van Dongen, thanks!